requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic==1.10.7
openai==0.27.8
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
python-dotenv==1.0.0
sqlalchemy==2.0.15
alembic==1.11.1
psycopg2-binary==2.9.6 